requirements.txt
requests
packaging
gitpython
